# SMS-Onboarding-Unified/api-specification.yaml
openapi: 3.0.3
info:
  title: SMS Onboarding Portal API
  description: |
    RESTful API for the SMS Onboarding Portal that manages vessel onboarding, equipment documentation,
    critical parts management, and integration with the maintenance portal.
  version: 1.0.0
  contact:
    name: SMS Support Team
    email: support@sms-portal.com
  license:
    name: Proprietary
servers:
  - url: https://api.sms-onboarding.com/v1
    description: Production server
  - url: https://staging-api.sms-onboarding.com/v1
    description: Staging server
  - url: http://localhost:4000/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /vessels:
    get:
      tags:
        - Vessels
      summary: List vessels
      operationId: listVessels
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/SortQuery'
        - name: companyId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VesselStatus'
        - name: onboardingStatus
          in: query
          schema:
            $ref: '#/components/schemas/OnboardingStatus'
      responses:
        '200':
          description: Vessels retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Vessels
      summary: Create a new vessel
      operationId: createVessel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVesselRequest'
      responses:
        '201':
          description: Vessel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vessel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /vessels/{vesselId}:
    get:
      tags:
        - Vessels
      summary: Get vessel by ID
      operationId: getVessel
      parameters:
        - $ref: '#/components/parameters/VesselId'
      responses:
        '200':
          description: Vessel retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vessel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Vessels
      summary: Update vessel
      operationId: updateVessel
      parameters:
        - $ref: '#/components/parameters/VesselId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVesselRequest'
      responses:
        '200':
          description: Vessel updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vessel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Vessels
      summary: Delete vessel
      operationId: deleteVessel
      parameters:
        - $ref: '#/components/parameters/VesselId'
      responses:
        '204':
          description: Vessel deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    webhookAuth:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
...
    VesselId:
      name: vesselId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    SearchQuery:
      name: search
      in: query
      schema:
        type: string
    SortQuery:
      name: sort
      in: query
      schema:
        type: string
        pattern: '^[a-zA-Z_]+:(asc|desc)$'
  schemas:
...
    CreateVesselRequest:
      type: object
      required:
        - companyId
        - name
        - imoNumber
        - flag
        - type
        - yearBuilt
        - grossTonnage
      properties:
        companyId:
          type: string
          format: uuid
        name:
          type: string
        imoNumber:
          type: string
          pattern: '^[0-9]{7}$'
        flag:
          type: string
        type:
          $ref: '#/components/schemas/VesselType'
        yearBuilt:
          type: integer
          minimum: 1900
        grossTonnage:
          type: number
        deadWeight:
          type: number
        length:
          type: number
        beam:
          type: number
        draft:
          type: number
        mainEngine:
          type: string
        auxiliaryEngines:
          type: array
          items:
            type: string
        classification:
          type: string
    UpdateVesselRequest:
      type: object
      properties:
        name:
          type: string
        flag:
          type: string
        type:
          $ref: '#/components/schemas/VesselType'
        status:
          $ref: '#/components/schemas/VesselStatus'
        mainEngine:
          type: string
        auxiliaryEngines:
          type: array
          items:
            type: string
        classification:
          type: string
...
    Vessel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        name:
          type: string
        imoNumber:
          type: string
        flag:
          type: string
        type:
          $ref: '#/components/schemas/VesselType'
        yearBuilt:
          type: integer
        grossTonnage:
          type: number
        deadWeight:
          type: number
        length:
          type: number
        beam:
          type: number
        draft:
          type: number
        mainEngine:
          type: string
        auxiliaryEngines:
          type: array
          items:
            type: string
        classification:
          type: string
        status:
          $ref: '#/components/schemas/VesselStatus'
        onboardingStatus:
          $ref: '#/components/schemas/OnboardingStatus'
        onboardingProgress:
          type: integer
          minimum: 0
          maximum: 100
        qualityScore:
          type: integer
          minimum: 0
          maximum: 100
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
...
    VesselListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vessel'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # Enums
...
    VesselType:
      type: string
      enum:
        - BULK_CARRIER
        - CONTAINER
        - TANKER
        - GENERAL_CARGO
        - RORO
        - PASSENGER
        - OFFSHORE
        - OTHER
    VesselStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - MAINTENANCE
        - DECOMMISSIONED
    OnboardingStatus:
      type: string
      enum:
        - NOT_STARTED
        - TOKEN_GENERATED
        - IN_PROGRESS
        - PENDING_REVIEW
        - APPROVED
        - REJECTED
        - COMPLETED
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...